+++ src/types/index.d.ts
@@ -26,5 +26,6 @@ export type * from './plugins/backup.d.ts';
 export type * from './plugins/cleanup.d.ts';
 export type * from './plugins/lokijs.d.ts';
 export type * from './plugins/dexie.d.ts';
+export type * from './plugins/reactivity.d.ts';
 export type * from './plugins/update.d.ts';
 export type * from './plugins/crdt.d.ts';